+++ ACP3/Modules/ACP3/Feeds/DependencyInjection/FeedAvailabilityCompilerPass.php
@@ -1,0 +1,33 @@
+<?php
+/**
+ * Copyright (c) 2016 by the ACP3 Developers.
+ * See the LICENCE file at the top-level module directory for licencing details.
+ */
+
+namespace ACP3\Modules\ACP3\Feeds\DependencyInjection;
+
+
+use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
+use Symfony\Component\DependencyInjection\ContainerBuilder;
+use Symfony\Component\DependencyInjection\Reference;
+
+class FeedAvailabilityCompilerPass implements CompilerPassInterface
+{
+    /**
+     * You can modify the container here before it is dumped to PHP code.
+     *
+     * @param ContainerBuilder $container
+     */
+    public function process(ContainerBuilder $container)
+    {
+        $definition = $container->findDefinition('feeds.utility.available_feeds_registrar');
+        $plugins = $container->findTaggedServiceIds('feeds.available_module');
+
+        foreach ($plugins as $serviceId => $tags) {
+            $definition->addMethodCall(
+                'registerModule',
+                [new Reference($serviceId)]
+            );
+        }
+    }
+}