+++ ACP3/Modules/ACP3/Files/Model/Repository/FilesRepository.php
@@ -120,7 +120,7 @@ class FilesRepository extends Core\Model\Repository\AbstractRepository
     {
         $period = ' AND ' . $this->getPublicationPeriod();
         return $this->db->fetchAll(
-            'SELECT id, title, `text` FROM ' . $this->getTableName() . ' WHERE MATCH (' . $fields . ') AGAINST (' . $this->db->getConnection()->quote($searchTerm) . ' IN BOOLEAN MODE)' . $period . ' ORDER BY `start` ' . $sortDirection . ', `end` ' . $sortDirection . ', id ' . $sortDirection,
+            'SELECT id, title, `text` FROM ' . $this->getTableName() . ' WHERE MATCH(' . $fields . ') AGAINST (' . $this->db->getConnection()->quote($searchTerm) . ' IN BOOLEAN MODE)' . $period . ' ORDER BY `start` ' . $sortDirection . ', `end` ' . $sortDirection . ', id ' . $sortDirection,
             ['time' => $time]
         );
     }