+++ ACP3/Modules/ACP3/Seo/Core/Pagination.php
@@ -6,11 +6,11 @@
 
 namespace ACP3\Modules\ACP3\Seo\Core;
 
+use ACP3\Core\Breadcrumb\Title;
 use ACP3\Core\Controller\AreaEnum;
 use ACP3\Core\Http\RequestInterface;
 use ACP3\Core\I18n\Translator;
 use ACP3\Core\Router\RouterInterface;
-use ACP3\Modules\ACP3\Seo\Core\Breadcrumb\Title;
 use ACP3\Modules\ACP3\Seo\Helper\MetaStatements;
 use ACP3\Modules\ACP3\Users\Model\UserModel;
 
@@ -27,13 +27,12 @@ class Pagination extends \ACP3\Core\Pagination
 
     /**
      * Pagination constructor.
-     *
-     * @param \ACP3\Modules\ACP3\Users\Model\UserModel $user
-     * @param \ACP3\Modules\ACP3\Seo\Core\Breadcrumb\Title $title
-     * @param \ACP3\Core\I18n\Translator $translator
-     * @param \ACP3\Core\Http\RequestInterface $request
-     * @param \ACP3\Core\Router\RouterInterface $router
-     * @param \ACP3\Modules\ACP3\Seo\Helper\MetaStatements $metaStatements
+     * @param UserModel $user
+     * @param Title $title
+     * @param Translator $translator
+     * @param RequestInterface $request
+     * @param RouterInterface $router
+     * @param MetaStatements $metaStatements
      */
     public function __construct(
         UserModel $user,
@@ -43,7 +42,6 @@ class Pagination extends \ACP3\Core\Pagination
         RouterInterface $router,
         MetaStatements $metaStatements
     ) {
-
         parent::__construct($user, $title, $translator, $request, $router);
 
         $this->metaStatements = $metaStatements;